liquid staking refactoring relayer and requested changes from review
