+++ code/parachain/frame/liquid-staking/relayer/src/main.rs
@@ -1,7 +1,11 @@
-use grandpa_client_primitives::parachain_header_storage_key;
-use grandpa_prover::GrandpaProver;
-
 use hyperspace_core::substrate::DefaultConfig as PolkadotConfig;
+use jsonrpsee::ws_client::WsClientBuilder;
+use sp_core::storage::StorageKey;
+use std::{str::FromStr, sync::Arc};
+use subxt::{
+	tx::{PairSigner, Payload},
+	OnlineClient, SubstrateConfig,
+};
 
 cfg_if::cfg_if! {
 	if #[cfg(feature = "composable")] {
@@ -29,22 +33,9 @@ cfg_if::cfg_if! {
 	}
 }
 
-use futures_util::stream::StreamExt;
-use jsonrpsee::{async_client::Client, tracing::log, ws_client::WsClientBuilder};
-use sp_core::{storage::StorageKey, Pair};
-use sp_keyring::AccountKeyring;
-use std::{io::Bytes, str::FromStr, sync::Arc, time::Duration};
-use subxt::{
-	config::Header,
-	dynamic::Value,
-	ext::scale_value::Composite,
-	rpc::types::StorageChangeSet,
-	tx::{PairSigner, Payload},
-	utils::AccountId32,
-	Config, OnlineClient, SubstrateConfig,
-};
+mod config;
 
-struct SovereignSubAccount {
+pub struct SovereignSubAccount {
 	pub address: String,
 	pub storage_key: StorageKey,
 	pub derivative_index: u16,
@@ -52,92 +43,28 @@ struct SovereignSubAccount {
 
 #[tokio::main]
 async fn main() {
-	//RELAY_HOST=wss://kusama-rpc.polkadot.io:443 PARA_HOST=ws://127.0.0.1:8000 cargo run -p
-	// lsd-relayer
+	/*
+		SEED="some twelve words" RELAY_HOST=ws://127.0.0.1:8000 PARA_HOST=ws://127.0.0.1:8001 SLEEP_TIME_MIN=60 cargo run -p lsd-relayer --features composable
+	*/
 
-	cfg_if::cfg_if! {
-		if #[cfg(feature = "composable")] {
-			println!("{}", "Composable");
-			let sovereign_account_id = "13YMK2ecbyxtm4cmFs31PqzWmQ7gWVboJSmXbcA56DB94xB9";
-			let sovereign_account_id_index_0 = "12x6QU4c9eRPxJMATFsRNFiZTMK5QgZkdZFFeu2QDKn4TR82";
-			let sovereign_account_id_index_1 = "1461Z7Bm1bwQpz1PuYMQ8phj9bRpxNU7ZYsb7aXQRAUuNecG";
-			let sovereign_account_id_index_2 = "15ySsNFkAhswdn9hSKkzoK7LhmJrj8bgyUZQAiM7Df9JpBUH";
-			let sovereign_account_id_index_3 = "15s3DuzMeftBH7YdHykwPDUd2DBxdNbiyqgDfZDA3i5eRwUW";
-			let sovereign_account_id_index_4 = "12uNvUSK39SDbHbqWuMhFdw2hHySrkbenVrHdS678fkj9BBb";
-			let sovereign_account_id_index_5 = "14tDkT3U93Pc1wLrHEjfYuhPPnFpMwDr7o8phPCTwTRj5wfE";
+	let sub_accounts = config::get_config();
 
-			let p0 = "5f3e4907f716ac89b6347d15ececedca422adb579f1dbf4f3886c5cfa3bb8cc456d08aea5b028f73699523ae21709a815640ec97748f5b5da9a2298e830e8971df7908861e1710b957fe06f0703bca7d";
-			let p1 = "5f3e4907f716ac89b6347d15ececedca422adb579f1dbf4f3886c5cfa3bb8cc491af1d8906a21795a98d84506b4216828886ca7474c66c027a9dc5d73901481568d551d01f13d0eb3bd36dd20ed2f13e";
-			let p2 = "5f3e4907f716ac89b6347d15ececedca422adb579f1dbf4f3886c5cfa3bb8cc4bc407118eef848de28bad224cc60c8fedbfdc1d87d5f65f94602d8e31d7426de7b7d0bb4a8e9223b72bb39231737377c";
-			let p3 = "5f3e4907f716ac89b6347d15ececedca422adb579f1dbf4f3886c5cfa3bb8cc43802b39b0fa95ce7e37486987631cef3d71aafa64aa8d5d9164071a14106cdba71fe4b8fc08de39a7606c8d16f5b20f8";
-			let p4 = "5f3e4907f716ac89b6347d15ececedca422adb579f1dbf4f3886c5cfa3bb8cc4b8ac2e6f27924e06455d8e5c75b6a775542eca539ca4e92c4b4e0496bef934595597645e21eb29b5ce94bc38b7e45181";
-			let p5 = "5f3e4907f716ac89b6347d15ececedca422adb579f1dbf4f3886c5cfa3bb8cc4232647f67d6923193db6aedafe0d5ea3abc522231610ec74f6391785db08c68f3c07c1bb45e2193b413f66926dec5072";
-		}
-		else{
-			println!("{}", "Kusama");
-			let sovereign_account_id = "F7fq1imhk2xwqqiQZav65eYKuzxSWWaF3HQp4fXkx6Sb3jY";
-			let sovereign_account_id_index_0 = "5xi8ChU4B7KQhpu3VcAhiPHJwCB7ZRBGHrtkkWEc59dufxud";
-			let sovereign_account_id_index_1 = "5vWm2EwDzaAx9m1zA4cn6L2kGpSqEodXCSboU3sC2L1Bap89";
-			let sovereign_account_id_index_2 = "5yDkX5PDf3HrS3DCfJCQNmXbGhXS5expcsQRuHF9PHijnBP8";
-			let sovereign_account_id_index_3 = "5uGqg72hYJwNbZsc8yUoosP6qKnaM9YmSLny97BiBWaFjVuF";
-			let sovereign_account_id_index_4 = "5twVfsa9HMayg9ewDaoGvBaxjsUY5CoWCdAyMkyJxVfJZzhL";
-			let sovereign_account_id_index_5 = "5xyCUmCz6X6Vt7BwqjD5TJrVztHa44SPbTsDbRFHtvNKoowG";
+	let seed_phrase = std::env::var("SEED").unwrap_or_else(|_| {
+		"private sentence hip meadow place say issue winner express edge royal aerobic".to_string()
+	});
 
-			let p0 = "5f3e4907f716ac89b6347d15ececedca422adb579f1dbf4f3886c5cfa3bb8cc4ca5bb2d842b7645d5be486494c94ae9eb6e44e3cc5ecf446f64b424c4df5e6c8f131eea4643d888b6520601991cab3ed";
-			let p1 = "5f3e4907f716ac89b6347d15ececedca422adb579f1dbf4f3886c5cfa3bb8cc46d27fba198fb3980b80b1cb2808485cb55c0d70b182f5eda4b3b431877942e51eed1e9f67605e3ff9796c37f990aa768";
-			let p2 = "5f3e4907f716ac89b6347d15ececedca422adb579f1dbf4f3886c5cfa3bb8cc4b8c748966862268009d367885803cc8fcd7cad008951936eedb5647e6fb1fcaabf026f5a8383909a18a9e460bbda98d5";
-			let p3 = "5f3e4907f716ac89b6347d15ececedca422adb579f1dbf4f3886c5cfa3bb8cc4ab8199cef4c3bcb677ed773063a7a59b1ee6d56d0cfb1c6fb2420245037c7e81f7502017ce171081faa51060c1893da2";
-			let p4 = "5f3e4907f716ac89b6347d15ececedca422adb579f1dbf4f3886c5cfa3bb8cc495c7694204423897be092f9d03b39e281025d4e00dc0e913ac982b3367b9503cd14cf62850f2c0e9bccf39dcfa780043";
-			let p5 = "5f3e4907f716ac89b6347d15ececedca422adb579f1dbf4f3886c5cfa3bb8cc436bc383c1b32aa2650599816d58630a2c263625eb7c0a0fc84845b51c699c67437e9fb698d2087cb7af53a0bfc8281cb";
-		}
-	}
+	let sleep_time_min = std::env::var("SLEEP_TIME_MIN").unwrap_or_else(|_| "60".to_string());
+	let sleep_time_min = sleep_time_min.parse::<u64>().unwrap_or_else(|_| 60);
 
-	let s0 = StorageKey(hex::decode(p0).expect("Failed to decode hex string p0"));
-	let s1 = StorageKey(hex::decode(p1).expect("Failed to decode hex string p1"));
-	let s2 = StorageKey(hex::decode(p2).expect("Failed to decode hex string p2"));
-	let s3 = StorageKey(hex::decode(p3).expect("Failed to decode hex string p3"));
-	let s4 = StorageKey(hex::decode(p4).expect("Failed to decode hex string p4"));
-	let s5 = StorageKey(hex::decode(p5).expect("Failed to decode hex string p5"));
-
-	let sub_accounts = vec![
-		SovereignSubAccount {
-			address: sovereign_account_id_index_0.to_string(),
-			storage_key: s0.clone(),
-			derivative_index: 0,
-		},
-		SovereignSubAccount {
-			address: sovereign_account_id_index_1.to_string(),
-			storage_key: s1.clone(),
-			derivative_index: 1,
-		},
-		SovereignSubAccount {
-			address: sovereign_account_id_index_2.to_string(),
-			storage_key: s2.clone(),
-			derivative_index: 2,
-		},
-		SovereignSubAccount {
-			address: sovereign_account_id_index_3.to_string(),
-			storage_key: s3.clone(),
-			derivative_index: 3,
-		},
-		SovereignSubAccount {
-			address: sovereign_account_id_index_4.to_string(),
-			storage_key: s4.clone(),
-			derivative_index: 4,
-		},
-		SovereignSubAccount {
-			address: sovereign_account_id_index_5.to_string(),
-			storage_key: s5.clone(),
-			derivative_index: 5,
-		},
-	];
 	let relay = std::env::var("RELAY_HOST")
 		.unwrap_or_else(|_| "wss://kusama-rpc.polkadot.io:443".to_string());
-	let para = std::env::var("PARA_HOST").unwrap_or_else(|_| "ws://127.0.0.1:8000".to_string());
+	let para = std::env::var("PARA_HOST")
+		.unwrap_or_else(|_| "wss://rpc.composablenodes.tech:443".to_string());
 
-	// let relay_ws_url = format!("wss://{relay}:443");
-	// let relay_ws_url = format!("ws://127.0.0.1:8001");
-	// let para_ws_url = format!("ws://127.0.0.1:8000");
+	println!("Configuration:");
+	println!("relay: {}", relay);
+	println!("para: {}", para);
+	println!("sleep_time_min: {}", sleep_time_min);
 
 	let relay_ws_url = relay.as_str();
 	let para_ws_url = para.as_str();
@@ -146,20 +73,16 @@ async fn main() {
 	let relay_client = OnlineClient::<PolkadotConfig>::from_rpc_client(relay_ws_client.clone())
 		.await
 		.unwrap();
-	let para_ws_client = Arc::new(WsClientBuilder::default().build(para_ws_url).await.unwrap());
-	let para_client = OnlineClient::<LsdConfig>::from_rpc_client(para_ws_client.clone())
-		.await
-		.unwrap();
+	let para_api = OnlineClient::<subxt::SubstrateConfig>::from_url(para_ws_url).await.unwrap();
 
-	while true {
+	loop {
 		for i in &sub_accounts {
 			let keys = vec![i.storage_key.as_ref()];
 
-			let api = OnlineClient::<subxt::SubstrateConfig>::from_url(para_ws_url).await.unwrap();
-			let block_hash_para = api.rpc().block_hash(None).await.unwrap().unwrap();
+			let block_hash_para = para_api.rpc().block_hash(None).await.unwrap().unwrap();
 			let validation_data =
 				parachain_subxt::api::storage().pallet_liquid_staking().validation_data();
-			let Some(validation_data) = api.storage().at(block_hash_para).fetch(&validation_data).await.unwrap()
+			let Some(validation_data) = para_api.storage().at(block_hash_para).fetch(&validation_data).await.expect("PalletLiquidStaking pallet should be available in the runtime")
 			else {
 				println!("validation data: {} not found", i.address);
 				continue;
@@ -167,7 +90,8 @@ async fn main() {
 
 			use crate::parachain_subxt::api::runtime_types::polkadot_primitives::v4::PersistedValidationData;
 
-			let validation_data = PersistedValidationData::try_from(validation_data).unwrap();
+			let validation_data = PersistedValidationData::try_from(validation_data)
+				.expect("Failed to decode PersistedValidationData");
 			// println!("validation_data: {:?}", validation_data);
 
 			let block_hash = relay_client
@@ -190,65 +114,58 @@ async fn main() {
 			assert!(state_proof.len() > 0);
 
 			use subxt::utils::AccountId32;
-			let account_id = AccountId32::from_str(&i.address).unwrap();
+			let account_id =
+				AccountId32::from_str(&i.address).expect("Failed to decode AccountId32");
 			let staking = relaychain::api::storage().staking().ledger(account_id);
 
-			let Some(ledger) = relay_client.storage().at(block_hash).fetch(&staking).await.unwrap()
+			let Some(ledger) = relay_client.storage().at(block_hash).fetch(&staking).await.expect("Staking pallet should be available in the runtime")
 			else {
 				println!("ledger: {} not found", i.address);
 				continue;
 			};
 
-			let sl =
+			let relaychain_staking_ledger =
 				relaychain::api::runtime_types::pallet_staking::StakingLedger::try_from(ledger)
-					.unwrap();
-			// println!("sl: {:?}", sl);
+					.expect("Failed to decode StakingLedger");
 
 			let mut unlocking = vec![];
-			for chunk in sl.unlocking.0.iter() {
+			for chunk in relaychain_staking_ledger.unlocking.0.iter() {
 				let e = UnlockChunk { value: chunk.value, era: chunk.era };
 				unlocking.push(e);
 			}
 
 			let mut claimed_rewards = vec![];
 
-			for claimed_reward in sl.claimed_rewards.0.iter() {
+			for claimed_reward in relaychain_staking_ledger.claimed_rewards.0.iter() {
 				let e = claimed_reward.clone();
 				claimed_rewards.push(e);
 			}
 
-			let xxx = StakingLedger::<AccountId32, u128> {
-				stash: AccountId32::from_str(&i.address).unwrap(),
-				total: sl.total,
-				active: sl.active,
+			let para_input_staking_ledger = StakingLedger::<AccountId32, u128> {
+				stash: AccountId32::from_str(&i.address).expect("Failed to decode AccountId32"),
+				total: relaychain_staking_ledger.total,
+				active: relaychain_staking_ledger.active,
 				unlocking,
 				claimed_rewards,
 			};
 			let tx_set_staking_ledger = parachain_subxt::api::tx()
 				.pallet_liquid_staking()
-				.set_staking_ledger(i.derivative_index, xxx, state_proof);
-			let tx_value =
-				parachain_subxt::api::tx().pallet_liquid_staking().initiate_exchange_rate();
+				.set_staking_ledger(i.derivative_index, para_input_staking_ledger, state_proof);
 
 			use subxt::ext::sp_core::Pair;
 			//test wallet for lsd testing 5DPqUqEfnp3buHaqiVnPt8ryykJEQRgdqAjbscnrZG2qDADa
-			let key = sp_keyring::sr25519::sr25519::Pair::from_string(
-				&"private sentence hip meadow place say issue winner express edge royal aerobic",
-				None,
-			)
-			.expect("secret");
+			let key = sp_keyring::sr25519::sr25519::Pair::from_string(&seed_phrase, None)
+				.expect("secret");
 			let signer: PairSigner<SubstrateConfig, sp_keyring::sr25519::sr25519::Pair> =
 				PairSigner::new(key.clone());
 
-			sign_and_submit_staking_ledger_update(&api, tx_set_staking_ledger, signer).await;
+			sign_and_submit_staking_ledger_update(&para_api, tx_set_staking_ledger, signer).await;
 			tokio::time::sleep(std::time::Duration::from_secs(10)).await;
 		}
 
-		//sleep 5 hours before next submit of ledger proof to lsd pallet on composable
-		tokio::time::sleep(std::time::Duration::from_secs(60 * 60 * 1)).await;
+		//sleep N hours before next submit of ledger proof to lsd pallet on composable
+		tokio::time::sleep(std::time::Duration::from_secs(60 * sleep_time_min)).await;
 	}
-
-	todo!();
 }
 
 async fn sign_and_submit_staking_ledger_update(
@@ -272,12 +189,3 @@ async fn sign_and_submit_staking_ledger_update(
 		}
 	}
 }
-
-// let event = relay_client.events();
-// let stream = relay_client.blocks().subscribe_finalized().await.unwrap()
-// .filter_map(|block| async {
-//     let block = block.ok().unwrap();
-//     let hash = block.hash();
-//     let events = event.at(hash).await.ok().unwrap();
-//     Some(events)
-// });
